requirements.txt
xmltodict
requests
sqlalchemy==2.0.0b2
python-telegram-bot==20.0a4 # pip install python-telegram-bot -U --pre